requirements.txt

fastapi
uvicorn
pyngrok
nest_asyncio
face_recognition
dlib
pillow
numpy
